--- a/arbor/include/arbor/spike_event.hpp
+++ b/arbor/include/arbor/spike_event.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <arbor/arb_types.hpp>
+
 #include <iosfwd>
 #include <tuple>
 #include <vector>
@@ -13,9 +15,12 @@ namespace arb {
 // Events delivered to targets on cells with a cell group.
 
 struct spike_event {
-    cell_lid_type target;
-    time_type time;
-    float weight;
+    cell_lid_type target = -1;
+    float weight = 0;
+    time_type time = -1;
+
+    spike_event() = default;
+    constexpr spike_event(cell_lid_type tgt, time_type t, arb_weight_type w) noexcept: target(tgt), weight(w), time(t) {}
 
     friend bool operator==(const spike_event& l, const spike_event& r) {
         return l.target==r.target && l.time==r.time && l.weight==r.weight;