--- a/arbor/include/arbor/spike_event.hpp
+++ b/arbor/include/arbor/spike_event.hpp
@@ -9,6 +9,7 @@
 #include <arbor/export.hpp>
 #include <arbor/serdes.hpp>
 #include <arbor/common_types.hpp>
+#include <arbor/util/lexcmp_def.hpp>
 
 namespace arb {
 
@@ -22,16 +23,10 @@ struct spike_event {
     spike_event() = default;
     constexpr spike_event(cell_lid_type tgt, time_type t, arb_weight_type w) noexcept: target(tgt), weight(w), time(t) {}
 
-    friend bool operator==(const spike_event& l, const spike_event& r) {
-        return l.target==r.target && l.time==r.time && l.weight==r.weight;
-    }
-
-    friend bool operator<(const spike_event& l, const spike_event& r) {
-        return std::tie(l.time, l.target, l.weight) < std::tie(r.time, r.target, r.weight);
-    }
-
     ARB_SERDES_ENABLE(spike_event, target, time, weight);
 };
+
+ARB_DEFINE_LEXICOGRAPHIC_ORDERING(spike_event,(a.time,a.target,a.weight),(b.time,b.target,b.weight))
 
 using pse_vector = std::vector<spike_event>;