@thorstenhater 2023-09-19
Spack fails on matplotlib deprecation `style=seaborn`

@boeschf We get these spurious (?) failures on daint, is there a way to re-run the jobs?